Change power.cpp to deal properly with kspace neutrinos, and with faking baryons to be part of the DM.
Also change initialise.cpp to keep track of the masses in each component
self-consistently with the new scheme.

OmegaBaryon is now used to specify the *physical* baryon fraction, ie,
the fraction which is fed to CAMB. If there are gas particles in the
glassfile, this will be used to find their mass, otherwise this amount
of baryons will be added to the dark matter.

There is a minor subtlety with neutrinos; in the early universe these
become relativistic and thus redshift differently, so their masses
relative to the rest of the DM change. CAMB knows how to deal with this,
so for fake neutrinos we will use CAMB's total transfer function column
and subtract off baryons if necessary.

For particle neutrinos we will still be wrong at high redshift, but we
the evolution will be wrong anyway.