Only include neutrinos as matter in the growth function if they are not free-streaming on scales 1/4 of the box.
If they are free-streaming, they do not cluster,
so this would overestimate the growth factor and thus the velocities.

This is not totally accurate if they are still free-streaming on
scales larger than the smallest resolved, but I think for those masses
or redshifts it is not very important. A warning is added anyway.